refactor src/helpers into relevant module
