--- api/src/helpers/index.ts
@@ -1,1 +1,0 @@
-export * as token from './token';
